Separate fetchMock into server.
