+++ src/webclient/main/services/authentication/index.js
@@ -1,79 +1,19 @@
 import VerificationError from '../../../models/errors/verification-error';
 import AuthorizationError from '../../../models/errors/authorization-error';
-import fetchMock from '../../../../../node_modules/fetch-mock/esm/client.js';
+
 import UserData from '../../../models/user-data';
+import MockServer from '../mock-server';
 
 /**
  * Implements login and registration methods logic.
  */
 export default class AuthenticationService {
   /**
-   * @typedef User
-   *  @param {string} login - users login.
-   *  @param {string} password - users password.
-   */
-
-  /**
-   * Registered users.
-   * @type {{User}}
-   */
-  users = {
-    admin: 'Admin123456',
-  };
-
-  /**
    * Creates object that intercepts requests for the server.
    */
-  constructor() {
-    console.log('Current users');
-    this.printUsers();
-
-    fetchMock
-      .post('/login', ((url, request) => {
-        const userData = new UserData(request.body.login, request.body.password);
-        if (this.isUserRegistered(userData)) {
-          return {
-            status: 200,
-            body: {token: `${userData.login}-token`},
-          };
-        } else {
-          return 401;
-        }
-      }));
-
-    fetchMock
-      .post('/registration', ((url, request) => {
-        const userData = new UserData(request.body.login, request.body.password);
-        if (!this.isLoginRegistered(userData) && userData.password.length >= 10) {
-          console.log(`New user ${userData.login} has been registered`);
-          this.users[userData.login.toLowerCase()] = userData.password;
-          return 200;
-        } else {
-          if (this.isLoginRegistered(userData)) {
-            return {
-              status: 401,
-              body: {
-                error: new AuthorizationError('User with this login already exists.'),
-              },
-            };
-          } else {
-            let errors = [];
-            errors.push(new VerificationError('password', 'Password should be longer than 10 characters.'));
-            return {
-              status: 422,
-              body: {
-                errors,
-              },
-            };
-          }
-        }
-      }));
-  }
-
-
-  printUsers() {
-    for (let p in this.users) {
-      console.log(p);
+  constructor(devMod) {
+    if (devMod) {
+      new MockServer();
     }
   }
 
@@ -146,44 +86,18 @@ export default class AuthenticationService {
     }));
   }
 
-  /**
-   * Checks if user with this login and password is registered.
-   * @param {UserData} userData - instance of {@link UserData}.
-   * @return {boolean} if user is registered returns true if it's not returns false.
-   */
-  isUserRegistered(userData) {
-    const login = userData.login.toLowerCase();
-    const password = userData.password;
-    return this.users.hasOwnProperty(login) && this.users[login] === password;
-  }
-
-  isLoginRegistered(userData) {
-    const login = userData.login.toLowerCase();
-    console.log(`login ${login} is ${this.users.hasOwnProperty(login)}`);
-    return this.users.hasOwnProperty(login);
-  }
-
   getItems() {
     return new Promise((resolve, reject) => {
-      const items = [
-        {
-          type: 'folder',
-          config: {name: 'Documents', itemsAmount: '2'},
-        },
-        {
-          type: 'folder',
-          config: {name: 'Images', itemsAmount: '2'},
-        },
-        {
-          type: 'folder',
-          config: {name: 'Videos', itemsAmount: '1'},
-        },
-        {
-          type: 'file',
-          config: {name: 'test.txt', mimeType: 'text', size: '20KB'},
-        },
-      ];
-      resolve(items);
+      fetch('/get-items', {
+        method: 'GET',
+      }).then(response => {
+        debugger;
+        const result = response.json();
+        result.then(items => {
+          console.log(items);
+          resolve(items);
+        });
+      });
     });
   }
 }
