+++ src/webclient/main/services/state-manager/mutators/item-loading-error-mutator/index.js
@@ -1,0 +1,13 @@
+import Mutator from '../mutator.js';
+
+export default class ItemLoadingErrorMutator extends Mutator {
+  constructor(error) {
+    super();
+    this.error = error;
+  }
+
+
+  apply(state) {
+    state.error = this.error;
+  }
+}
