+++ src/webclient/main/services/state-manager/mutators/item-loading-mutator/index.js
@@ -1,0 +1,13 @@
+import Mutator from '../mutator.js';
+
+export default class ItemLoadingMutator extends Mutator {
+  constructor(isLoading) {
+    super();
+    this.isLoading = isLoading;
+  }
+
+
+  apply(state) {
+    state.isLoading = this.isLoading;
+  }
+}
